refactor: room에서 사용하는 타입 이동
- types로 이동시킴
