+++ packages/backend/src/domains/room/entity/room.entity.ts
@@ -1,18 +1,5 @@
+import { GAME_MODE, RoomData } from "../../../types/room";
 import { Player } from "./player.entitiy";
-
-export type GAME_MODE = "CatchMind" | "Garticphone" | "";
-
-export interface RoomData {
-  roomId: string;
-  players?: Player[];
-  host?: string;
-  state?: boolean;
-  gameMode?: GAME_MODE;
-  totalRound?: number;
-  roundTime?: number;
-  goalScore?: number;
-  maxPlayer?: number;
-}
 
 export class Room {
   roomId: string;
